+++ bee-common/bee-packable/src/packable/integer.rs
@@ -1,7 +1,7 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{coerce::*, error::UnpackError, packer::Packer, unpacker::Unpacker, Packable};
+use crate::{error::UnpackError, packer::Packer, unpacker::Unpacker, Packable, UnpackErrorExt};
 
 use core::convert::Infallible;
