+++ bee-message/src/message.rs
@@ -7,7 +7,7 @@ use crate::{
     MessageId, MessageUnpackError, ValidationError,
 };
 
-use bee_packable::{coerce::*, Packable, Packer, UnpackError, Unpacker};
+use bee_packable::{Packable, Packer, UnpackError, UnpackErrorExt, Unpacker};
 
 use crypto::{
     hashes::{blake2b::Blake2b256, Digest},
