+++ bee-message/src/metadata.rs
@@ -3,7 +3,7 @@
 
 use crate::{payload::fpc::Opinion, MessageUnpackError};
 
-use bee_packable::{coerce::*, Packable, Packer, UnpackError, Unpacker};
+use bee_packable::{Packable, Packer, UnpackError, UnpackErrorExt, Unpacker};
 
 use bitflags::bitflags;
